+++ src/connection.h
@@ -1,20 +1,10 @@
 #pragma once
 
-#include <QColorDialog>
-#include <QCoreApplication>
 #include <QDialog>
-#include <QJsonArray>
-#include <QJsonDocument>
-#include <QJsonObject>
-#include <QJsonValue>
-#include <QMessageBox>
-#include <QSqlDatabase>
-#include <QSqlError>
-#include <QSqlQuery>
+#include <QString>
 
-namespace Ui {
-    class Connection;
-}
+class ConnectionBackend;
+class QQuickWidget;
 
 class Connection : public QDialog
 {
@@ -23,20 +13,8 @@ class Connection : public QDialog
 public:
     explicit Connection(QString selectedHost, QWidget* parent = nullptr);
     ~Connection();
-    void saveConnection();
-
-private slots:
-    void on_buttonCancel_clicked();
-
-    void on_buttonSave_clicked();
-
-    void on_buttonRemove_clicked();
-
-    void on_buttonConnect_clicked();
-
-    void on_dial_valueChanged(int value);
 
 private:
-    Ui::Connection* ui;
+    ConnectionBackend* backend;
+    QQuickWidget* quickWidget;
 };
-